fix: changing the valor field from float to double in transacao.java

+++ src/main/java/com/example/itau/transacao/model/Transacao.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Transacao {
     private Integer id;
 
-    private float valor;
+    private double valor;
 
     private OffsetDateTime dataHora;
 }
